refactor: split ChatRepository.kt into two classes MessageRepository.kt and ChatRepository.kt
