+++ core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/di/repositoryModule.kt
@@ -2,9 +2,11 @@ package net.thechance.mena.core_chat.data.di
 
 import net.thechance.mena.core_chat.data.repository.ChatRepositoryImpl
 import net.thechance.mena.core_chat.data.repository.ContactsRepositoryImpl
+import net.thechance.mena.core_chat.data.repository.MessageRepositoryImpl
 import net.thechance.mena.core_chat.data.repository.UserRepositoryImpl
 import net.thechance.mena.core_chat.domain.repository.ChatRepository
 import net.thechance.mena.core_chat.domain.repository.ContactsRepository
+import net.thechance.mena.core_chat.domain.repository.MessageRepository
 import net.thechance.mena.core_chat.domain.repository.UserRepository
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
@@ -23,13 +25,19 @@ internal val repositoryModule = module {
     single<ChatRepository> {
         ChatRepositoryImpl(
             client = get(named(CHAT_CLIENT)),
-            json = get(named(CHAT_JSON)),
             webSocketManager = get(),
-            messageDao = get(),
             imageDownloader = get()
         )
     }
+   single<MessageRepository>{
+       MessageRepositoryImpl(
+           client = get(named(CHAT_CLIENT)),
+           webSocketManager = get(),
+           messageDao = get(),
+           json = get(named(CHAT_JSON))
+       )
 
+   }
     single<UserRepository> {
         UserRepositoryImpl(
             client = get(named(CHAT_CLIENT)),
