+++ core_chat_data/src/commonTest/kotlin/net/thechance/mena/core_chat/data/chat/ChatRepositoryImplTest.kt
@@ -77,56 +77,8 @@ class ChatRepositoryImplTest {
         repository = createChatRepository(
             httpClient = httpClient,
             webSocketManager = webSocketManager,
-            messageDao = messageDao,
             imageDownloader = imageDownloader
         )
-    }
-
-    @Test
-    fun `should return messages when loadMessages is successful`() = runTest {
-        httpClient = createHttpClient(
-            chatHistoryResponse = { defaultChatHistoryResponse() }
-        )
-        repository = createChatRepository(
-            httpClient = httpClient,
-            webSocketManager = webSocketManager,
-            messageDao = messageDao,
-            imageDownloader = imageDownloader
-        )
-
-        val result = repository.loadMessages(chatId)
-
-        assertThat(result).isNotEmpty()
-    }
-
-    @Test
-    fun `should delete message from local database when deleteMessage is called`() = runTest {
-        val message = createMessage(
-            senderId = userId,
-            chatId = chatId,
-        )
-        everySuspend { messageDao.deleteMessage(any()) } returns Unit
-
-        repository.deleteMessage(message)
-
-        verifySuspend { messageDao.deleteMessage(message.id.toString()) }
-    }
-
-    @Test
-    fun `should throw NotFoundException when loadMessages returns error`() = runTest {
-        httpClient = createHttpClient(
-            chatHistoryResponse = { mockErrorPagedResponse<MessageDto>(HttpStatusCode.NotFound) }
-        )
-        repository = createChatRepository(
-            httpClient = httpClient,
-            webSocketManager = webSocketManager,
-            messageDao = messageDao,
-            imageDownloader = imageDownloader
-        )
-
-        assertFailsWith<NotFoundException> {
-            repository.loadMessages(chatId)
-        }
     }
 
 
@@ -136,7 +88,6 @@ class ChatRepositoryImplTest {
         repository = createChatRepository(
             httpClient = httpClient,
             webSocketManager = webSocketManager,
-            messageDao = messageDao,
             imageDownloader = imageDownloader
         )
 
@@ -153,7 +104,6 @@ class ChatRepositoryImplTest {
         repository = createChatRepository(
             httpClient = httpClient,
             webSocketManager = webSocketManager,
-            messageDao = messageDao,
             imageDownloader = imageDownloader
         )
 
@@ -179,7 +129,6 @@ class ChatRepositoryImplTest {
         repository = createChatRepository(
             httpClient = httpClient,
             webSocketManager = webSocketManager,
-            messageDao = messageDao,
             imageDownloader = imageDownloader
         )
 
@@ -201,7 +150,6 @@ class ChatRepositoryImplTest {
         repository = createChatRepository(
             httpClient = httpClient,
             webSocketManager = webSocketManager,
-            messageDao = messageDao,
             imageDownloader = imageDownloader
         )
 
@@ -218,7 +166,6 @@ class ChatRepositoryImplTest {
         repository = createChatRepository(
             httpClient = httpClient,
             webSocketManager = webSocketManager,
-            messageDao = messageDao,
             imageDownloader = imageDownloader
         )
 
@@ -237,7 +184,6 @@ class ChatRepositoryImplTest {
         repository = createChatRepository(
             httpClient = httpClient,
             webSocketManager = webSocketManager,
-            messageDao = messageDao,
             imageDownloader = imageDownloader
         )
 
@@ -249,103 +195,6 @@ class ChatRepositoryImplTest {
         }
     }
 
-    @Test
-    fun `should send message successfully when websocket is connected`() = runTest {
-        every { webSocketManager.isConnected() } returns true
-        everySuspend { webSocketManager.sendTextFrame(any(), any()) } returns Unit
-        everySuspend { messageDao.insertMessage(any()) } returns Unit
-        everySuspend { messageDao.deleteMessage(any()) } returns Unit
-
-        val message = createMessage(
-            senderId = userId,
-            chatId = chatId,
-        )
-
-        repository.sendMessage(message)
-
-        verifySuspend {
-            webSocketManager.sendTextFrame(
-                destination = "/app/chat.privateMessage",
-                payload = any()
-            )
-        }
-    }
-
-    @Test
-    fun `should throw SendMessageFailedException when websocket is not connected`() = runTest {
-        every { webSocketManager.isConnected() } returns false
-        everySuspend { messageDao.insertMessage(any()) } returns Unit
-        everySuspend { messageDao.updateMessageStatus(any(), any()) } returns Unit
-
-        val message = createMessage(
-            senderId = userId,
-            chatId = chatId,
-        )
-
-        assertFailsWith<SendMessageFailedException> {
-            repository.sendMessage(message)
-        }
-    }
-
-    @Test
-    fun `should disconnect websocket when disconnect is called`() = runTest {
-        everySuspend { webSocketManager.disconnect() } returns Unit
-        repository.disconnect()
-        verifySuspend { webSocketManager.disconnect() }
-    }
-
-    @Test
-    fun `should observe read messages`() = runTest {
-        val flow = repository.observeReadMessages()
-        assertThat(flow).isNotNull()
-    }
-
-    @Test
-    fun `should return local messages from database when getLocalMessages is called`() = runTest {
-        val message1 = createMessage(senderId = userId, chatId = chatId)
-        val message2 = createMessage(senderId = userId, chatId = chatId)
-        val messageEntities = listOf(
-            message1.toLocalDto(),
-            message2.toLocalDto()
-        )
-
-        everySuspend { messageDao.getMessagesByChat(chatId.toString()) } returns flowOf(messageEntities)
-
-        val result = repository.getLocalMessages(chatId).first()
-
-
-        assertThat(result).isNotEmpty()
-        assertThat(result.size).isEqualTo(2)
-        verifySuspend { messageDao.getMessagesByChat(chatId.toString()) }
-    }
-
-    @Test
-    fun `should return empty list when no local messages exist for chat`() = runTest {
-        everySuspend { messageDao.getMessagesByChat(chatId.toString()) } returns flowOf(emptyList())
-
-        val result = repository.getLocalMessages(chatId).first()
-
-        assertThat(result.isEmpty()).isTrue()
-        verifySuspend { messageDao.getMessagesByChat(chatId.toString()) }
-    }
-
-    @Test
-    fun `should return flow when getMessages is called and websocket is connected`() = runTest {
-        every { webSocketManager.isConnected() } returns true
-        everySuspend { authRepository.getAccessToken() } returns "test-token"
-        everySuspend { webSocketManager.connect(any()) } returns Unit
-        everySuspend { webSocketManager.subscribe(any()) } returns Unit
-        everySuspend { webSocketManager.sendTextFrame(any(), any()) } returns Unit
-        every { webSocketManager.incomingMessages } returns MutableSharedFlow<String>().apply {
-            tryEmit(
-                "test-message"
-            )
-        }
-
-        val flow = repository.getMessages(chatId)
-
-        assertThat(flow).isNotNull()
-    }
 
     @Test
     fun `downloadImage should call imageDownloader and run successfully when downloadImageToGallery return true`() =
@@ -384,7 +233,6 @@ class ChatRepositoryImplTest {
         repository = createChatRepository(
             httpClient = httpClient,
             webSocketManager = webSocketManager,
-            messageDao = messageDao,
             imageDownloader = imageDownloader
         )
 
@@ -404,7 +252,6 @@ class ChatRepositoryImplTest {
         repository = createChatRepository(
             httpClient = httpClient,
             webSocketManager = webSocketManager,
-            messageDao = messageDao,
             imageDownloader = imageDownloader
         )
 
@@ -413,74 +260,8 @@ class ChatRepositoryImplTest {
         }
     }
 
-    @Test
-    fun `should send image message successfully when websocket connected and images uploaded`() = runTest {
-        every { webSocketManager.isConnected() } returns true
-        everySuspend { messageDao.insertMessage(any()) } returns Unit
-        everySuspend { messageDao.updateMessageImages(any(), any()) } returns Unit
-        everySuspend { messageDao.deleteMessage(any()) } returns Unit
-
-        httpClient = createHttpClient(
-            imagesResponse = { defaultUploadImagesResponse() }
-        )
-        repository = createChatRepository(
-            httpClient = httpClient,
-            webSocketManager = webSocketManager,
-            messageDao = messageDao,
-            imageDownloader = imageDownloader
-        )
-
-        val byteArrays = listOf(ByteArray(10), ByteArray(20))
-        val message = createMessage(
-            senderId = userId,
-            chatId = chatId,
-            content = MessageContent.Images(ImagesSource.Local(byteArrays))
-        )
-
-        repository.sendMessage(message)
-
-        verifySuspend { messageDao.insertMessage(any()) }
-        verifySuspend { messageDao.updateMessageImages(any(), any()) }
-        verifySuspend { messageDao.deleteMessage(any()) }
-    }
-
-
-    @Test
-    fun `should mark message as FAILED when image upload throws exception`() = runTest {
-        every { webSocketManager.isConnected() } returns true
-        everySuspend { messageDao.insertMessage(any()) } returns Unit
-        everySuspend { messageDao.updateMessageStatus(any(), any()) } returns Unit
-
-        // Simulate failed upload
-        httpClient = createHttpClient(
-            imagesResponse = { respondError(HttpStatusCode.InternalServerError) }
-        )
-        repository = createChatRepository(
-            httpClient = httpClient,
-            webSocketManager = webSocketManager,
-            messageDao = messageDao,
-            imageDownloader = imageDownloader
-        )
-
-        val byteArrays = listOf(ByteArray(10))
-        val message = createMessage(
-            senderId = userId,
-            chatId = chatId,
-            content = MessageContent.Images(ImagesSource.Local(byteArrays))
-        )
-
-        assertFailsWith<SendMessageFailedException> {
-            repository.sendMessage(message)
-        }
-
-        verifySuspend { messageDao.updateMessageStatus(any(), MessageLocalDto.MessageStatus.FAILED) }
-    }
-
-
     private companion object {
-        private val chatId = Uuid.random()
         private val userId = Uuid.random()
-
         const val IMAGE_URL = "http://test.com/image.jpg"
     }
