+++ core_chat_data/src/commonTest/kotlin/net/thechance/mena/core_chat/data/chat/MessageRepositoryImplTest.kt
@@ -1,0 +1,274 @@
+@file:OptIn(ExperimentalUuidApi::class, ExperimentalTime::class)
+
+package net.thechance.mena.core_chat.data.chat
+
+import assertk.assertThat
+import assertk.assertions.isEqualTo
+import assertk.assertions.isNotEmpty
+import assertk.assertions.isNotNull
+import assertk.assertions.isTrue
+import dev.mokkery.answering.returns
+import dev.mokkery.every
+import dev.mokkery.everySuspend
+import dev.mokkery.matcher.any
+import dev.mokkery.mock
+import dev.mokkery.verifySuspend
+import io.ktor.client.HttpClient
+import io.ktor.client.engine.mock.respondError
+import io.ktor.http.HttpStatusCode
+import kotlinx.coroutines.flow.MutableSharedFlow
+import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.flowOf
+import kotlinx.coroutines.test.runTest
+import net.thechance.mena.core_chat.data.contacts.fakes.createMessage
+import net.thechance.mena.core_chat.data.createHttpClient
+import net.thechance.mena.core_chat.data.createMessageRepository
+import net.thechance.mena.core_chat.data.defaultChatHistoryResponse
+import net.thechance.mena.core_chat.data.defaultUploadImagesResponse
+import net.thechance.mena.core_chat.data.mockErrorPagedResponse
+import net.thechance.mena.core_chat.data.repository.MessageRepositoryImpl
+import net.thechance.mena.core_chat.data.source.local.database.MessageDao
+import net.thechance.mena.core_chat.data.source.local.database.MessageLocalDto
+import net.thechance.mena.core_chat.data.source.remote.dto.MessageDto
+import net.thechance.mena.core_chat.data.source.remote.mapper.toLocalDto
+import net.thechance.mena.core_chat.data.source.remote.network.WebSocketManager
+import net.thechance.mena.core_chat.domain.entity.ImagesSource
+import net.thechance.mena.core_chat.domain.entity.MessageContent
+import net.thechance.mena.core_chat.domain.exception.NotFoundException
+import net.thechance.mena.core_chat.domain.exception.SendMessageFailedException
+import kotlin.test.BeforeTest
+import kotlin.test.Test
+import kotlin.test.assertFailsWith
+import kotlin.time.ExperimentalTime
+import kotlin.uuid.ExperimentalUuidApi
+import kotlin.uuid.Uuid
+
+class MessageRepositoryImplTest {
+
+    private lateinit var httpClient: HttpClient
+    private lateinit var repository: MessageRepositoryImpl
+    private lateinit var webSocketManager: WebSocketManager
+    private lateinit var messageDao: MessageDao
+
+    @BeforeTest
+    fun setUp() {
+        httpClient = createHttpClient()
+        webSocketManager = mock<WebSocketManager>()
+        messageDao = mock<MessageDao>()
+
+        repository = createMessageRepository(
+            webSocketManager = webSocketManager,
+            messageDao = messageDao,
+            httpClient = httpClient
+        )
+    }
+
+    @Test
+    fun `should return messages when loadMessages is successful`() = runTest {
+        httpClient = createHttpClient(
+            chatHistoryResponse = { defaultChatHistoryResponse() }
+        )
+        repository = createMessageRepository(
+            webSocketManager = mock<WebSocketManager>(),
+            messageDao = mock<MessageDao>(),
+            httpClient = httpClient
+        )
+
+        val result = repository.loadMessages(chatId)
+
+        assertThat(result).isNotEmpty()
+    }
+
+    @Test
+    fun `should delete message from local database when deleteMessage is called`() = runTest {
+        val message = createMessage(
+            senderId = userId,
+            chatId = chatId,
+        )
+        everySuspend { messageDao.deleteMessage(any()) } returns Unit
+
+        repository.deleteMessage(message)
+
+        verifySuspend { messageDao.deleteMessage(message.id.toString()) }
+    }
+
+    @Test
+    fun `should throw NotFoundException when loadMessages returns error`() = runTest {
+        httpClient = createHttpClient(
+            chatHistoryResponse = { mockErrorPagedResponse<MessageDto>(HttpStatusCode.NotFound) }
+        )
+        repository = createMessageRepository(
+            httpClient = httpClient,
+            webSocketManager = webSocketManager,
+            messageDao = messageDao,
+        )
+
+        assertFailsWith<NotFoundException> {
+            repository.loadMessages(chatId)
+        }
+    }
+
+    @Test
+    fun `should send message successfully when websocket is connected`() = runTest {
+        every { webSocketManager.isConnected() } returns true
+        everySuspend { webSocketManager.sendTextFrame(any(), any()) } returns Unit
+        everySuspend { messageDao.insertMessage(any()) } returns Unit
+        everySuspend { messageDao.deleteMessage(any()) } returns Unit
+
+        val message = createMessage(
+            senderId = userId,
+            chatId = chatId,
+        )
+
+        repository.sendMessage(message)
+
+        verifySuspend {
+            webSocketManager.sendTextFrame(
+                destination = "/app/chat.privateMessage",
+                payload = any()
+            )
+        }
+    }
+
+    @Test
+    fun `should throw SendMessageFailedException when websocket is not connected`() = runTest {
+        every { webSocketManager.isConnected() } returns false
+        everySuspend { messageDao.insertMessage(any()) } returns Unit
+        everySuspend { messageDao.updateMessageStatus(any(), any()) } returns Unit
+
+        val message = createMessage(
+            senderId = userId,
+            chatId = chatId,
+        )
+
+        assertFailsWith<SendMessageFailedException> {
+            repository.sendMessage(message)
+        }
+    }
+
+
+    @Test
+    fun `should observe read messages`() = runTest {
+        val flow = repository.observeReadMessages()
+        assertThat(flow).isNotNull()
+    }
+
+    @Test
+    fun `should return local messages from database when getLocalMessages is called`() = runTest {
+        val message1 = createMessage(senderId = userId, chatId = chatId)
+        val message2 = createMessage(senderId = userId, chatId = chatId)
+        val messageEntities = listOf(
+            message1.toLocalDto(),
+            message2.toLocalDto()
+        )
+
+        everySuspend { messageDao.getMessagesByChat(chatId.toString()) } returns flowOf(
+            messageEntities
+        )
+
+        val result = repository.getLocalMessages(chatId).first()
+
+
+        assertThat(result).isNotEmpty()
+        assertThat(result.size).isEqualTo(2)
+        verifySuspend { messageDao.getMessagesByChat(chatId.toString()) }
+    }
+
+    @Test
+    fun `should return empty list when no local messages exist for chat`() = runTest {
+        everySuspend { messageDao.getMessagesByChat(chatId.toString()) } returns flowOf(emptyList())
+
+        val result = repository.getLocalMessages(chatId).first()
+
+        assertThat(result.isEmpty()).isTrue()
+        verifySuspend { messageDao.getMessagesByChat(chatId.toString()) }
+    }
+
+    @Test
+    fun `should return flow when getMessages is called and websocket is connected`() = runTest {
+        every { webSocketManager.isConnected() } returns true
+        everySuspend { webSocketManager.connect(any()) } returns Unit
+        everySuspend { webSocketManager.subscribe(any()) } returns Unit
+        everySuspend { webSocketManager.sendTextFrame(any(), any()) } returns Unit
+        every { webSocketManager.incomingMessages } returns MutableSharedFlow<String>().apply {
+            tryEmit(
+                "test-message"
+            )
+        }
+
+        val flow = repository.getMessages(chatId)
+
+        assertThat(flow).isNotNull()
+    }
+
+    @Test
+    fun `should send image message successfully when websocket connected and images uploaded`() =
+        runTest {
+            every { webSocketManager.isConnected() } returns true
+            everySuspend { messageDao.insertMessage(any()) } returns Unit
+            everySuspend { messageDao.updateMessageImages(any(), any()) } returns Unit
+            everySuspend { messageDao.deleteMessage(any()) } returns Unit
+
+            httpClient = createHttpClient(
+                imagesResponse = { defaultUploadImagesResponse() }
+            )
+            repository = createMessageRepository(
+                httpClient = httpClient,
+                webSocketManager = webSocketManager,
+                messageDao = messageDao
+            )
+
+            val byteArrays = listOf(ByteArray(10), ByteArray(20))
+            val message = createMessage(
+                senderId = userId,
+                chatId = chatId,
+                content = MessageContent.Images(ImagesSource.Local(byteArrays))
+            )
+
+            repository.sendMessage(message)
+
+            verifySuspend { messageDao.insertMessage(any()) }
+            verifySuspend { messageDao.updateMessageImages(any(), any()) }
+            verifySuspend { messageDao.deleteMessage(any()) }
+        }
+
+    @Test
+    fun `should mark message as FAILED when image upload throws exception`() = runTest {
+        every { webSocketManager.isConnected() } returns true
+        everySuspend { messageDao.insertMessage(any()) } returns Unit
+        everySuspend { messageDao.updateMessageStatus(any(), any()) } returns Unit
+
+        // Simulate failed upload
+        httpClient = createHttpClient(
+            imagesResponse = { respondError(HttpStatusCode.InternalServerError) }
+        )
+        repository = createMessageRepository(
+            httpClient = httpClient,
+            webSocketManager = webSocketManager,
+            messageDao = messageDao
+        )
+
+        val byteArrays = listOf(ByteArray(10))
+        val message = createMessage(
+            senderId = userId,
+            chatId = chatId,
+            content = MessageContent.Images(ImagesSource.Local(byteArrays))
+        )
+
+        assertFailsWith<SendMessageFailedException> {
+            repository.sendMessage(message)
+        }
+
+        verifySuspend {
+            messageDao.updateMessageStatus(
+                any(),
+                MessageLocalDto.MessageStatus.FAILED
+            )
+        }
+    }
+
+    private companion object {
+        private val chatId = Uuid.random()
+        private val userId = Uuid.random()
+    }
+}
